refactor: move Thunderbird event handler and content script registration to an appropriate Class
